Move LegalEntityDiscountingProvider
