+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/bond/BlackBondFutureOptionMarginedProductPricer.java
@@ -9,7 +9,6 @@ import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.market.product.bond.BondFutureOptionSensitivity;
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.pricer.impl.option.BlackFormulaRepository;
-import com.opengamma.strata.pricer.rate.LegalEntityDiscountingProvider;
 import com.opengamma.strata.product.bond.ResolvedBondFuture;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOption;
 import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
