+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/bond/BlackBondFutureOptionMarginedTradePricer.java
@@ -8,7 +8,6 @@ package com.opengamma.strata.pricer.bond;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.market.product.bond.BondFutureOptionSensitivity;
-import com.opengamma.strata.pricer.rate.LegalEntityDiscountingProvider;
 import com.opengamma.strata.product.bond.ResolvedBondFuture;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOption;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOptionTrade;
