+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/bond/BondFutureOptionMarginedProductPricer.java
@@ -6,7 +6,6 @@
 package com.opengamma.strata.pricer.bond;
 
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
-import com.opengamma.strata.pricer.rate.LegalEntityDiscountingProvider;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOption;
 import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
