+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/bond/BondFutureOptionMarginedTradePricer.java
@@ -10,7 +10,6 @@ import java.time.LocalDate;
 import com.opengamma.strata.basics.currency.CurrencyAmount;
 import com.opengamma.strata.basics.currency.MultiCurrencyAmount;
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
-import com.opengamma.strata.pricer.rate.LegalEntityDiscountingProvider;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOption;
 import com.opengamma.strata.product.bond.ResolvedBondFutureOptionTrade;
 import com.opengamma.strata.product.common.FutureOptionPremiumStyle;
