+++ modules/pricer/src/main/java/com/opengamma/strata/pricer/bond/DiscountingBondFutureProductPricer.java
@@ -11,7 +11,6 @@ import com.opengamma.strata.collect.ArgChecker;
 import com.opengamma.strata.market.product.CompoundedRateType;
 import com.opengamma.strata.market.sensitivity.PointSensitivities;
 import com.opengamma.strata.market.sensitivity.PointSensitivityBuilder;
-import com.opengamma.strata.pricer.rate.LegalEntityDiscountingProvider;
 import com.opengamma.strata.product.bond.ResolvedBondFuture;
 import com.opengamma.strata.product.bond.ResolvedFixedCouponBond;
