+++ modules/pricer/src/test/java/com/opengamma/strata/pricer/bond/LegalEntityDiscountingProviderTest.java
@@ -3,7 +3,7 @@
  *
  * Please see distribution for license.
  */
-package com.opengamma.strata.pricer.rate;
+package com.opengamma.strata.pricer.bond;
 
 import static com.opengamma.strata.basics.currency.Currency.GBP;
 import static com.opengamma.strata.basics.currency.Currency.USD;
